update reduce sum/min/max/mean//prod case
